Pin constraints: re-organize pin constraint files for configurability.
Couple the pin constraint files with the variant of design blocks.
This should allow us to customize cape I/Os to best suit a specific
cape. It also allows to easily include or excluse a block from the
design. For example, not include cape, M.2, MIPI-CSI or high speed
interface if not required and tight on FPGA resources.