CI: Use git hash as version number for custom gateware CI.
BeagleV-Fire uses the PolarFire-Soc IAP feature to reprogram the
SoC-FPGA from Linux. The IAP mode used only actually programs the device
if the Libero design version is different from the one in the currently
programmed FPGA design.
This change to the gateware builder uses the gateware repo's git hash to
generate a Libero design version number which will be different with each
push to a forked gateware repository. This results in the bitstream
generated by the CI to successfully reprogram the SoC-FPGA after each
change to the gateware source code.